+++ bff/db/sqlc/tx_create_person.go
@@ -35,7 +35,7 @@ func (store *SQLStore) CreatePersonTx(ctx context.Context, arg CreatePersonTxPar
 
 		err = q.CloneProviders(ctx, CloneProvidersParams{
 			Creator:  arg.Creator,
-			PersonID: result.Person.ID,
+			PersonID: int64(result.Person.ID),
 		})
 		return err
 	})
